add parser for defaults and nones, thread flags through to Planner
