+++ ducktape/src/main/scala/io/github/arainko/ducktape/internal/Planner.scala
@@ -10,139 +10,15 @@ import scala.collection.immutable.VectorMap
 import scala.quoted.*
 import scala.util.boundary
 
-case class PlanFlags(source: SideSpecficFlags, dest: SideSpecficFlags) derives Debug {
-  def transition[A](
-    sourceStep: Step | Passthrough,
-    destStep: Step | Passthrough
-  )(using Quotes): PlanFlags = this.copy(source.transition(sourceStep), dest.transition(destStep))
-
-  inline def locally[A](inline f: PlanFlags ?=> A): A = f(using this)
-}
-
-object PlanFlags {
-  def current(using f: PlanFlags): f.type = f
-}
-
-enum Effect {
-  case Defaults, Nones
-}
-
-case class Flag(effect: Effect, kind: Kind) derives Debug
-
-// What to support:
-// * 'local' flags - i.e. ones that disappear in the next transition step once they reach their destination
-// * 'regional' flags - i.e. ones that stick around all the way down till they reach the leaf transformations
-// * 'type-specific' flags - like global, but only apply to a given type (can they also be local?)
-//
-//
-enum Kind derives Debug {
-  final def isLocal: Boolean =
-    this match
-      case Local                       => true
-      case Regional                    => false
-      case TypeSpecific(tpe, Local)    => true
-      case TypeSpecific(tpe, Regional) => false
-
-  case Local
-  case Regional
-  case TypeSpecific(tpe: Type[?], kind: Local.type | Regional.type)
-}
-
-case object Passthrough
-type Passthrough = Passthrough.type
-
-enum Step { self =>
-  case Element
-  case Field(name: String)
-  case TupleElement(index: Int)
-  case Case(tpe: Type[?])
-
-  final infix def =:=(that: Step)(using Quotes): Boolean =
-    (self, that) match {
-      case (Element, Element)                             => true
-      case (Field(selfName), Field(thatName))             => selfName == thatName
-      case (TupleElement(selfIdx), TupleElement(thatIdx)) => selfIdx == thatIdx
-      case (Case(selfTpe), Case(thatTpe))                 => selfTpe.repr =:= thatTpe.repr
-      case _                                              => false
-    }
-}
-
-object Step {
-  def fromPathSegment(segment: Path.Segment): Step =
-    segment match {
-      case Path.Segment.Field(tpe, name)         => Field(name)
-      case Path.Segment.TupleElement(tpe, index) => TupleElement(index)
-      case Path.Segment.Case(tpe)                => Case(tpe)
-      case Path.Segment.Element(tpe)             => Element
-    }
-}
-
-case class SideSpecficFlags(
-  outOfScope: Vector[(List[Step], Flag)],
-  inScope: Vector[Flag]
-) derives Debug {
-
-  def has(effect: Effect): Boolean = inScope.exists(_.effect == effect)
-
-  def transition(step: Step | Passthrough)(using Quotes): SideSpecficFlags = {
-    val (nextInScope, nextOutOfScope) = outOfScope.partitionMap { segmentsAndFlag =>
-      (step *: segmentsAndFlag) match {
-        case (_, Nil, flag) =>
-          Left(flag)
-        case (Passthrough, path, flag) =>
-          Right(Some(path, flag))
-        case (segment: Step, head :: Nil, flag) if head =:= segment =>
-          Left(flag)
-        case (segment: Step, head :: tail, flag) if head =:= segment =>
-          Right(Some((tail, flag)))
-        // prune these, it means this won't match next matches either (I thiiiiiiiiiiink?)
-        case (segment: Step, other, flag) =>
-          Right(None)
-      }
-
-    }
-
-    SideSpecficFlags(
-      nextOutOfScope.flatten,
-      nextInScope ++ inScope.filter(!_.kind.isLocal)
-    )
-  }
-}
-
-object SideSpecficFlags {
-  def current(using f: SideSpecficFlags): f.type = f
-
-  def create(outOfScope: Vector[(List[Step], Flag)],inScope: Vector[Flag]) = {
-    val (immediateInScope, outsideOfScope) = outOfScope.partitionMap { 
-      case (Nil, flag) => Left(flag)
-      case (other, flag) => Right(other -> flag)
-    }
-    SideSpecficFlags(outsideOfScope, immediateInScope ++ inScope)
-  }
-}
-
 private[ducktape] object Planner {
   import Structure.*
   private enum FallthroughUpcast {
     case Yes, No
   }
 
-  def between[F <: Fallible](source: Structure, dest: Structure)(using Quotes, Context.Of[F]) = {
+  def between[F <: Fallible](source: Structure, dest: Structure, flags: PlanFlags)(using Quotes, Context.Of[F]) = {
     given Depth = Depth.zero
-    given PlanFlags =
-      PlanFlags(
-        SideSpecficFlags.create(Vector.empty, Vector.empty),
-        SideSpecficFlags.create(
-          Vector(
-            // Nil -> Flag(Effect.Defaults, Kind.Local),
-            List(
-              // Step.Field("level1"),
-              // Step.Field("int")
-            ) -> Flag(Effect.Defaults, Kind.Regional)
-          ),
-          Vector.empty
-        )
-      )
+    given PlanFlags = flags
     recurse(source, dest)
   }
